refacotr: Cut of process workspace view
